requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
Flask-Cors
marshmallow-sqlalchemy
PyJWT
pymysql
numpy
scipy